requirements.txt
Babel==2.6.0
Django==2.0.8
django-phonenumber-field==2.0.0
phonenumberslite==8.9.10
pytz==2018.5
